Removing config_fields and refactoring config command
