+++ genai-perf/genai_perf/checkpoint/checkpoint.py
@@ -16,7 +16,8 @@ import json
 import os
 from dataclasses import dataclass
 
-from genai_perf.config.input.config_command import ConfigCommand, default_field
+from genai_perf.config.input.config_command import ConfigCommand
+from genai_perf.config.input.config_defaults import default_field
 from genai_perf.config.run.results import Results
 from genai_perf.exceptions import GenAIPerfException
